% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapro.R
\name{MMPO}
\alias{MMPO}
\title{MMPO}
\usage{
MMPO(p, is.onetail = TRUE, eff.sign)
}
\arguments{
\item{p}{A vector of p-values}

\item{is.onetail}{Logical. If set TRUE, p-values are combined without considering the direction of effect, and vice versa. Default: TRUE.}

\item{eff.sign}{A vector of signs of effect sizes. It works when is.onetail = FALSE}
}
\value{
A list of Combined p-value based on MMPO method, the indices of effecitve p-values and NA values.
}
\description{
Minimum Marginal P-value in joint order distribution
}
\examples{
MMPO(p=c(0.01, 0.02, 0.2, 0.5, 0.8), is.onetail=FALSE, eff.sign=c(1,1,1,1,1))
}
