% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapro.R
\name{MMPO}
\alias{MMPO}
\title{MMPO}
\usage{
MMPO(p, is.onetail = TRUE, eff.sign = NULL)
}
\arguments{
\item{p}{A vector of p-values}

\item{is.onetail}{Logical. If set TRUE, p-values are combined without considering the direction of effect, and vice versa. Default: TRUE.}

\item{eff.sign}{A vector of signs of effect sizes. It works when is.onetail = FALSE}
}
\value{
p : Combined p-value

optimal_rank : Optimal rank where minimum marginal p-value exists.

eff.p.idx : Index of effective p-values

MMP : Minimum marginal p-value

overall.eff.direction : The direction of combined effects.
}
\description{
Minimum Marginal P-value in joint order distribution
}
\examples{
MMPO(p=c(0.01, 0.02, 0.2, 0.5, 0.8), is.onetail=FALSE, eff.sign=c(1,1,1,1,1))
}
